refactor: move signCloudToken next to token.manager
`src/utils/` is a grab bag; token signing and verification now sit in the same
directory. Kept as a plain function rather than a TokenManager method: that
class signs credentials for our own surfaces and verifies inbound cloud tokens
against cloud's JWKS, and its constructor throws a bare Error when JWT_SECRET is
empty — which is the self-host case these five providers have to report cleanly.
